Ensure python3-hvac is installed for charms with encypt option
The referenced bug is essentially: make vault:secrets relation to vault
but keep the 'encrypt' option as False. In this case, the Context
handling code in charm-helpers is expecting python3-hvac to be
available, but it is only installed if the encrypt option is set to
True.  Hence the charm crashes.  This resolves that crash.

Note the related charm-helpers fix [1].

[1]: https://github.com/juju/charm-helpers/pull/431

Change-Id: I92773b7c1f48d456091062751e69581fabe4c5f3
Closes-bug: #1862085