Monitor Swift Object/Container/Account API availability
As per the bug report, it's not enough to simply monitor that the
appropriate process is alive; there have been instances of the process
being alive but the port/API being unavailable.

This patch adds monitoring for Object/Container/Account API
availability.

I've tested the fix in my small test environment and I can confirm
it's working.

For reference, the following branch/commit was used as a functional
test (later rejected with the aim of moving the checks to Mojo):
https://github.com/openstack-charmers/zaza-openstack-tests/pull/395

Change-Id: I60c5b74279f71ca8f8bc769c93af2eab1f59e002
Closes-Bug: #1854299